Summering av beløp i WriteToFileService endres fra int til long

+++ src/main/kotlin/no/nav/sokos/spk/mottak/util/FileParser.kt
@@ -121,7 +121,7 @@ object FileParser {
 
     fun createEndRecord(
         antallTransaksjon: Int,
-        sumBelop: Int,
+        sumBelop: Long,
     ): String {
         val stringBuilder = StringBuilder()
         stringBuilder
